backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.124.4
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
